+++ dataconv/types/many.go
@@ -1,4 +1,4 @@
-package util
+package types
 
 import (
 	"fmt"
